refactor(kernels): split speculative sampling into its own module (review)
Address @xiaguan's #534 review:
- Move gpu_verify_probs_into + gpu_spec_accept_into (and their tests) into
  ops/spec_sampling.rs; sampling.rs 1336 -> 1060 lines, so the speculative
  code no longer drives the file over the soft size limit.
- Extract shared C helper gather_softmax_renorm_flashinfer (gather + softmax
  + optional top-k/top-p renorm), used by both the verify path and the min_p
  sampling path; extract Rust prepare_sampling_params shared by the batched
  sampler and the verify path. min_p stays caller-side (sampler packs it,
  verify rejects it).
- Expose onehot_draft as a bool parameter on gpu_spec_accept_into instead of
  hardcoding 1, so the slice-2 verify path can pass a real drafter's probs.
- Guard vocab > 0 in gpu_spec_accept_into.
- Add spec_accept_partial_acceptance_matches_the_rate_law: a 512-row
  Bernoulli(0.7) convergence test for the p in (0,1) acceptance law.

Verified on H100: cargo test -p openinfer-kernels --lib (sampling + spec)
passes 11/11.
