+++ openinfer-kernels/src/ops.rs
@@ -15,6 +15,7 @@ mod linear;
 mod lora;
 mod norm;
 mod sampling;
+mod spec_sampling;
 
 pub use attention::{
     PrefillPagedPlan, SUPPORTED_GQA_GROUP_SIZES, dflash_qk_norm_rope_into,
@@ -70,9 +71,9 @@ pub use sampling::{
     BatchSamplingRow, BatchSamplingScratch, argmax, argmax_batch_bf16_into,
     argmax_batch_bf16_split_indexed_into, argmax_batch_bf16_split_partials_len, argmax_bf16_into,
     argmax_bf16_split_into, flashinfer_top1_batch_into, flashinfer_top1_row_states_bytes,
-    gpu_sample_batch_into, gpu_spec_accept_into, gpu_verify_probs_into, markov_step_argmax_into,
-    markov_step_argmax_partials_len,
+    gpu_sample_batch_into, markov_step_argmax_into, markov_step_argmax_partials_len,
 };
+pub use spec_sampling::{gpu_spec_accept_into, gpu_verify_probs_into};
 
 /// Calling thread's last FFI exception message, ready to append to an error;
 /// empty unless `result` is the -1 sentinel set by the C++ guard. Public for
